% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANdllogger.R
\name{cran.download.log}
\alias{cran.download.log}
\title{cran.download.log}
\usage{
cran.download.log(start.date,end.date, package.name)
}
\arguments{
\item{start.date}{Starting date is the date since user want to extract the package download history log.}

\item{end.date}{End date is the date till the user want to extract the package download history log.}

\item{package.name}{The name of the package.}
}
\value{
Registry of downloading history of the package for the specified dates
}
\description{
CRANdllogger is a R package which enables the R package developer to download CRAN downloading history in csv file format. The package downloads the complete downloading history for the duration specified as input, when triggered for the first time. Once the complete database is downloaded, the tools can work as standalone package, without any need of internet. The output file can be procured from the output.csv directory in the current working directory.
}
\details{
CRANdllogger
}
